+++ GoRogue/Components/ParentAware/ParentAwareComponentBase.cs
@@ -5,6 +5,28 @@ using JetBrains.Annotations;
 namespace GoRogue.Components.ParentAware
 {
     /// <summary>
+    ///  EventArguments used for the <see cref="ParentAwareComponentBase.Removed"/> event.
+    /// </summary>
+    /// <typeparam name="T">The type of the parent being passed.</typeparam>
+    [PublicAPI]
+    public class ParentAwareComponentRemovedEventArgs<T> : EventArgs
+    {
+        /// <summary>
+        /// The parent from which the object was detached.
+        /// </summary>
+        public readonly T OldParent;
+
+        /// <summary>
+        /// Constructor.
+        /// </summary>
+        /// <param name="oldParent">The parent from which the object was detached.</param>
+        public ParentAwareComponentRemovedEventArgs(T oldParent)
+        {
+            OldParent = oldParent;
+        }
+    }
+
+    /// <summary>
     /// Simple (and optional) base class for components attached to a class implementing
     /// <see cref="IObjectWithComponents"/>.  Adds useful events and some helper functions to allow performing
     /// type-checking of parent, or requiring that the object it's attached to has or does not have certain types of
@@ -21,7 +43,7 @@ namespace GoRogue.Components.ParentAware
         /// <summary>
         /// Fires when the component is unattached from an object
         /// </summary>
-        public event EventHandler? Removed;
+        public event EventHandler<ParentAwareComponentRemovedEventArgs<IObjectWithComponents>>? Removed;
 
         private IObjectWithComponents? _parent;
         /// <summary>
@@ -36,8 +58,10 @@ namespace GoRogue.Components.ParentAware
 
                 if (value == null)
                 {
+                    var oldValue = _parent;
                     _parent = value;
-                    Removed?.Invoke(this, EventArgs.Empty);
+                    // Null for oldValue is not possible because value == null AND value != _parent
+                    Removed?.Invoke(this, new ParentAwareComponentRemovedEventArgs<IObjectWithComponents>(oldValue!));
                 }
                 else
                 {
@@ -109,8 +133,20 @@ namespace GoRogue.Components.ParentAware
         }
 
         /// <summary>
+        /// Fires when the component is unattached from an object
+        /// </summary>
+        public new event EventHandler<ParentAwareComponentRemovedEventArgs<TParent>>? Removed;
+
+        /// <summary>
         /// Constructor.
         /// </summary>
-        public ParentAwareComponentBase() => Added += ParentTypeCheck<TParent>;
+        public ParentAwareComponentBase()
+        {
+            Added += ParentTypeCheck<TParent>;
+            base.Removed += OnRemoved;
+        }
+
+        private void OnRemoved(object? sender, ParentAwareComponentRemovedEventArgs<IObjectWithComponents> e)
+            => Removed?.Invoke(sender, new ParentAwareComponentRemovedEventArgs<TParent>((TParent)e.OldParent));
     }
 }